Move option parsing to files
